server/requirements.txt
alembic==1.11.1
fastapi[standard]
uvicorn[standard]
sqlalchemy==2.0.30
pydantic-settings==2.0.2
python-dateutil==2.8.2
psycopg2==2.9.9